+++ changelog.md
@@ -14,6 +14,7 @@ Codion Change Log
 - TabbedPaneBuilder.prepareComponent() added.
 - DefaultFilterTableSearchModel bug fixed, predicate now notifies when SET.
 - FilterTable, search field keyboard actions improved.
+- InputDialogBuilder.close(Predicate) added, replacing overloaded show() method.
 ### is.codion.framework.model
 - EntityEditModel.Settings moved to EntityEditor, related refactoring.
 - AbstractEntityTableModel.deleteSelected() removed, related refactoring.